Report on flaky tests separately
Read a file whose path has been provided as an ENV variable. That file
contains a list of flaky tests that should not be reported in the normal
fashion if they fail. Most reporters will mark them as skipped, but for
the JUnitXML reporter, we add a flaky_test attribute to the testCase xml
for the flaky tests.