Exit 1 when there is an uninitialized constant
We are currently exiting with 0 when there is an error such as
uninitialized constant. Buildkite and probably other CI systems will
then consider this status as successful and the suite will only fail
when the aggregate of the executed tests is checked.
With this change the individual steps will also fail.

This also changes the behavior of the before-suite failures which will
now start exiting with 1, see https://github.com/Shopify/ci-queue/pull/73/files.