Implement a release command
If a worker hard crash (e.g. OOMKilled), other workers will have to
wait for its reversed test to reach the timeout before they pick it up.

In many cases we could detect these crashes and timeout the dead worker
lease immediately.

We could even have the `run` command fork a worker and monitor it
to automatically release the reserved tests if it crashed.