Refactor test ACK to reduce the race condition window
Because the `ack` and the error registration are so far
apart, it's possible that the reporter exit successfully
even though one test failed.

The worker first ack that the test was ran, then report
the error on stdout and other places, then finally record
the error in redis.

This leave a relatively big window during which the reporter
might see that the queue is empty, but no errors is reported
yet.

To fix that we must ack and report the error atomically,
which require to move the ack inside the error reporter.

This commit doesn't fully close the race condition but make it
much smaller, and make it much easier to bundle the ack and
error registration in a single atomic lua script.

Co-Authored-By: Christian Bruckmayer <christian.bruckmayer@shopify.com>