Fix failure file dump when output has invalid UTF8
Previously, any invalid UTF-8 in test output would prevent the entire
JSON failure file from being dumped, because the Ruby JSON library
validates that its output is UTF-8.

This commit aims to fix this issue by encoding the output as UTF-8
before it gets to JSON so that the file can always be dumped.